Use JSON.stringify instead of .toString() to serialize Book objects
When the user clicks on the "View list of books" button all that was being shown was "[Object Object]" which does not show any fields of the book to the user - not being of much use. Using JSON.stringify() instead for now we can see the different fields stored for this book.